+++ apps/pipeline-modeling-tool/dummy_source.hpp
